Fix ownership of wp-contant folder, setting it to default (www-data) Workaround on docker-library/wordpress/issues/232 [https://github.com/docker-library/wordpress/issues/232]
More info:
[tianon commented on 30 Oct
2017](https://github.com/docker-library/wordpress/issues/232#issuecomment-340613133)
    WordPress expects to self-update (and IMO anyone who doesn't allow
it to do so these days is asking for trouble), so we simply recommend
that users instead make the entire install a directory and be done with
it. If you'd like a fresh install from an image to include
pre-installed themes or plugins, simply add them to the source
directory in /usr/src/wordpress (which is what gets copied into
/var/www/html at runtime).

Applying a VOLUME in this way is consistent with how other official
images handle their explicitly mutable state.

This definitely could do with some better documentation, which is over
at
https://github.com/docker-library/docs/blob/master/wordpress/content.md
if anyone wanted to take a stab at writing something.]
https://github.com/docker-library/wordpress/issues/232#issuecomment-340613133